@@ -1,19 +1,21 @@
+import 'dart:convert';
 import 'dart:io';
 
-import 'package:WOLapp/models/machine_definition.dart';
-import 'package:WOLapp/packet_sender.dart';
-import 'package:WOLapp/storage/machine_store.dart';
-import 'package:WOLapp/storage/sqlite_store.dart';
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
-import 'package:package_info/package_info.dart';
+import 'package:package_info_plus/package_info_plus.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:share_plus/share_plus.dart';
 import 'package:url_launcher/url_launcher.dart';
+import 'package:wol_app/models/machine_definition.dart';
+import 'package:wol_app/packet_sender.dart';
+import 'package:wol_app/storage/machine_store.dart';
+import 'package:wol_app/storage/sqlite_store.dart';
 
-final RegExp reMacAddress = RegExp(
-    r'^((([0-9A-F]{2}-){5}[0-9A-F]{2})|(([0-9A-F]{2}:){5}[0-9A-F]{2}))$',
-    caseSensitive: false);
+final RegExp reMacAddress = RegExp(r'^((([0-9A-F]{2}-){5}[0-9A-F]{2})|(([0-9A-F]{2}:){5}[0-9A-F]{2}))$', caseSensitive: false);
 
 final MachineStore machineStore = new SqliteMachineStore();
-PackageInfo packageInfo;
+PackageInfo? packageInfo;
 
 void main() async {
   runApp(WolApp());
@@ -35,14 +37,15 @@ class WolApp extends StatelessWidget {
 }
 
 class MachineSelectionList extends StatefulWidget {
-  MachineSelectionList({Key key}) : super(key: key);
+  MachineSelectionList({Key? key}) : super(key: key);
 
   @override
   _MachineSelectionListState createState() => _MachineSelectionListState();
 }
 
 class _MachineSelectionListState extends State<MachineSelectionList> {
-  List<MachineDefinition> _machineDefinitions;
+  List<MachineDefinition> _machineDefinitions = [];
+  bool _loaded = false;
 
   _MachineSelectionListState() {
     machineStore.loadMachines().then(_loadedFromDb);
@@ -51,15 +54,13 @@ class _MachineSelectionListState extends State<MachineSelectionList> {
   void _loadedFromDb(List<MachineDefinition> definitions) {
     setState(() {
       _machineDefinitions = definitions;
+      _loaded = true;
     });
   }
 
   void _editMachine(BuildContext context, MachineDefinition definition) {
     Navigator.of(context).push(PageRouteBuilder(
-      pageBuilder: (context, animation, secondaryAnimation) => MachineEditForm(
-          definition,
-          onSave: _saveModifiedDefinition,
-          onDelete: _deleteMachine),
+      pageBuilder: (context, animation, secondaryAnimation) => MachineEditForm(definition, onSave: _saveModifiedDefinition, onDelete: _deleteMachine),
       transitionsBuilder: (context, animation, secondaryAnimation, child) {
         final tween = Tween(begin: Offset(0.0, 1.0), end: Offset.zero);
         final offsetAnimation = animation.drive(tween);
@@ -109,43 +110,62 @@ class _MachineSelectionListState extends State<MachineSelectionList> {
                 title: Text('About'),
                 onTap: () {
                   Navigator.pop(context);
-                  showAboutDialog(
-                      context: context,
-                      applicationName: "WOL App",
-                      applicationVersion: packageInfo.version,
-                      children: <Widget>[
-                        TextButton(
-                          onPressed: () =>
-                              launch("https://github.com/mormegil-cz/wolapp"),
-                          child: Text("GitHub"),
-                        ),
-                      ]);
+                  showAboutDialog(context: context, applicationName: "WOL App", applicationVersion: packageInfo!.version, children: <Widget>[
+                    TextButton(
+                      onPressed: () => launch("https://github.com/mormegil-cz/wolapp"),
+                      child: Text("GitHub"),
+                    ),
+                  ]);
+                },
+              ),
+              ListTile(
+                title: Text('Export machines'),
+                enabled: _machineDefinitions.length > 0,
+                onTap: () async {
+                  Navigator.pop(context);
+                  final exportJson = await machineStore.exportMachinesJson();
+                  final tempDir = await getTemporaryDirectory();
+                  final jsonFile = File('${tempDir.path}/wolapp-export.json');
+                  await jsonFile.writeAsString(exportJson, mode: FileMode.writeOnly);
+                  await Share.shareFilesWithResult([jsonFile.path], mimeTypes: ['application/octet-stream'], subject: 'WOL App machine export');
+                  await jsonFile.delete();
+                },
+              ),
+              ListTile(
+                title: Text('Import machines'),
+                enabled: _machineDefinitions.length == 0,
+                onTap: () async {
+                  Navigator.pop(context);
+
+                  final pickedFile = await FilePicker.platform.pickFiles(withData: true);
+                  if (pickedFile != null) {
+                    final fileBytes = pickedFile.files.first.bytes!;
+                    final importJson = new Utf8Decoder().convert(fileBytes);
+                    await machineStore.importMachinesJson(importJson);
+                  }
                 },
               ),
             ],
           ),
         ),
-        body: _machineDefinitions == null
-            ? Center(child: CircularProgressIndicator())
-            : _machineDefinitions.isEmpty
+        body: _loaded
+            ? _machineDefinitions.isEmpty
                 ? Center(
                     child: Column(mainAxisSize: MainAxisSize.min, children: [
                       Text("No machines defined!"),
-                      Text("Press the button to add a first one!"),
+                      Text("Press the button to add a first one"),
+                      Text("or import machines in the menu!"),
                     ]),
                   )
                 : Scrollbar(
                     child: ListView(
                       padding: const EdgeInsets.symmetric(vertical: 8),
-                      children: _machineDefinitions
-                          .map((machine) => MachineListTile(
-                              machine: machine, onTap: _editMachine))
-                          .toList(growable: false),
+                      children: _machineDefinitions.map((machine) => MachineListTile(machine: machine, onTap: _editMachine)).toList(growable: false),
                     ),
-                  ),
+                  )
+            : Center(child: CircularProgressIndicator()),
         floatingActionButton: FloatingActionButton(
-          onPressed: () =>
-              {_editMachine(context, new MachineDefinition(machineIndex: -1))},
+          onPressed: () => {_editMachine(context, new MachineDefinition(machineIndex: -1))},
           tooltip: 'Add new machine',
           child: Icon(Icons.add),
         ),
@@ -154,31 +174,22 @@ class _MachineSelectionListState extends State<MachineSelectionList> {
 
 class MachineListTile extends StatelessWidget {
   final MachineDefinition machine;
-  final void Function(BuildContext buildContext, MachineDefinition machine)
-      onTap;
+  final void Function(BuildContext buildContext, MachineDefinition machine) onTap;
 
-  const MachineListTile({Key key, this.machine, this.onTap}) : super(key: key);
+  const MachineListTile({Key? key, required this.machine, required this.onTap}) : super(key: key);
 
   @override
   Widget build(BuildContext context) => ListTile(
-        leading: Icon(machine.getIcon(),
-            color:
-                machine.color ?? Theme.of(context).textTheme.bodyText2.color),
-        title: Text(machine.getCaption()),
+        leading: Icon(machine.getIcon(), color: machine.color ?? Theme.of(context).textTheme.bodyText2?.color),
+        title: Text(machine.getCaption()!),
         trailing: IconButton(
-          icon:
-              Icon(Icons.power_settings_new, semanticLabel: "Send wake packet"),
+          icon: Icon(Icons.power_settings_new, semanticLabel: "Send wake packet"),
           onPressed: () async {
             try {
-              await sendWakeUpPacket(machine.macAddress, machine.ipAddress,
-                  machine.port, machine.password);
-              ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                  content:
-                      Text("Wakeup packet for ${machine.getCaption()} sent")));
+              await sendWakeUpPacket(machine.macAddress!, machine.ipAddress!, machine.port!, machine.password);
+              ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text("Wakeup packet for ${machine.getCaption()} sent")));
             } catch (e) {
-              ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-                  content: Text("Error sending wakeup packet"),
-                  backgroundColor: Colors.redAccent));
+              ScaffoldMessenger.of(context).showSnackBar(SnackBar(content: Text("Error sending wakeup packet"), backgroundColor: Colors.redAccent));
             }
           },
         ),
@@ -193,11 +204,10 @@ class MachineEditForm extends StatefulWidget {
   final void Function(MachineDefinition definition) onSave;
   final void Function(int machineIndex) onDelete;
 
-  MachineEditForm(this.definition, {this.onSave, this.onDelete});
+  MachineEditForm(this.definition, {required this.onSave, required this.onDelete});
 
   @override
-  _MachineEditFormState createState() =>
-      _MachineEditFormState(definition, onSave: onSave, onDelete: onDelete);
+  _MachineEditFormState createState() => _MachineEditFormState(definition, onSave: onSave, onDelete: onDelete);
 }
 
 class _MachineEditFormState extends State<MachineEditForm> {
@@ -213,15 +223,12 @@ class _MachineEditFormState extends State<MachineEditForm> {
   final TextEditingController _ctrlPortNumber;
   final TextEditingController _ctrlPassword;
 
-  _MachineEditFormState(MachineDefinition definition,
-      {this.onSave, this.onDelete})
+  _MachineEditFormState(MachineDefinition definition, {required this.onSave, required this.onDelete})
       : _editedIndex = definition.machineIndex,
         _ctrlCaption = TextEditingController(text: definition.caption),
         _ctrlMacAddress = TextEditingController(text: definition.macAddress),
-        _ctrlIpAddress = TextEditingController(
-            text: definition.ipAddress ?? "255.255.255.255"),
-        _ctrlPortNumber =
-            TextEditingController(text: definition.port?.toString() ?? "9"),
+        _ctrlIpAddress = TextEditingController(text: definition.ipAddress ?? "255.255.255.255"),
+        _ctrlPortNumber = TextEditingController(text: definition.port?.toString() ?? "9"),
         _ctrlPassword = TextEditingController(text: definition.password);
 
   @override
@@ -237,9 +244,7 @@ class _MachineEditFormState extends State<MachineEditForm> {
   @override
   Widget build(BuildContext context) => Scaffold(
       appBar: AppBar(
-        title: Text(_editedIndex < 0
-            ? "WOLApp: Add new machine"
-            : "WOLApp: Edit machine"),
+        title: Text(_editedIndex < 0 ? "WOLApp: Add new machine" : "WOLApp: Edit machine"),
         actions: [
           if (_editedIndex >= 0)
             IconButton(
@@ -253,7 +258,7 @@ class _MachineEditFormState extends State<MachineEditForm> {
               icon: Icon(Icons.check),
               tooltip: "Save changes",
               onPressed: () {
-                if (_formKey.currentState.validate()) {
+                if (_formKey.currentState?.validate() ?? false) {
                   onSave(new MachineDefinition(
                     machineIndex: _editedIndex,
                     caption: _ctrlCaption.text,
@@ -315,14 +320,14 @@ class _MachineEditFormState extends State<MachineEditForm> {
             ),
           ])));
 
-  static String _validateMacAddress(String value) {
-    if (value.isEmpty) return 'This is a required field';
+  static String? _validateMacAddress(String? value) {
+    if (value == null || value.isEmpty) return 'This is a required field';
     if (!reMacAddress.hasMatch(value)) return 'Enter a valid MAC address';
     return null;
   }
 
-  static String _validateIpAddress(String value) {
-    if (value.isEmpty) return 'This is a required field';
+  static String? _validateIpAddress(String? value) {
+    if (value == null || value.isEmpty) return 'This is a required field';
     try {
       InternetAddress(value);
     } catch (e) {
@@ -331,12 +336,11 @@ class _MachineEditFormState extends State<MachineEditForm> {
     return null;
   }
 
-  static String _validatePortNumber(String value) {
-    if (value.isEmpty) return 'This is a required field';
+  static String? _validatePortNumber(String? value) {
+    if (value == null || value.isEmpty) return 'This is a required field';
     final parsedPort = int.tryParse(value);
     if (parsedPort == null) return 'Enter a valid port number';
-    if (parsedPort < 0 || parsedPort > 65535)
-      return 'Enter a valid port between 0 and 65535';
+    if (parsedPort < 0 || parsedPort > 65535) return 'Enter a valid port between 0 and 65535';
     return null;
   }
 }
